pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Markus Lang <markuslang@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11,<3.13"
uvicorn = "^0.32.0"
langchain = "^0.3.7"
langgraph = "^0.2.48"
fastapi = "^0.115.5"
langchain-openai = "^0.2.8"
langchain-community = "^0.3.7"
python-dotenv = "^1.0.1"
psycopg2 = "^2.9.10"
supabase = "^2.10.0"
pandas = "^2.2.3"



[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
isort = "^5.13.2"
black = "^24.10.0"
ruff = "^0.7.4"
pytest-asyncio = "^0.24.0"
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"